fix imports for extracted lively.components
