+++ ide/js/browser/index.js
@@ -5,12 +5,12 @@ import { morph, StyleSheet, show, Label, HorizontalLayout, GridLayout,
          DropDownList, config, Window } from "lively.morphic";
 import InputLine from "../../../text/input-line.js";
 import JSONEditorPlugin from "lively.morphic/ide/json/editor-plugin.js";
-import { HorizontalResizer } from "lively.morphic/components/resizers.js";
-import { Icon } from "lively.morphic/components/icons.js";
+import { HorizontalResizer } from "lively.components/resizers.js";
+import { Icon } from "lively.morphic/text/icons.js";
 import JavaScriptEditorPlugin from "../editor-plugin.js";
 import EvalBackendChooser from "../eval-backend-ui.js";
 import browserCommands from "./commands.js";
-import { Tree, TreeData } from "lively.morphic/components/tree.js"
+import { Tree, TreeData } from "lively.components/tree.js"
 
 import "mocha-es6/index.js";
 
@@ -1001,7 +1001,7 @@ export default class Browser extends Window {
 
       if (attempt > 0 || err instanceof SyntaxError)
         return sourceEditor.showError(err);
-      
+
       // try to reload the module, sometimes format changes (global => esm etc need a reload)
       let result = await this.reloadModule(false);
       sourceEditor.textString = content;
@@ -1025,7 +1025,7 @@ export default class Browser extends Window {
         resetEnv = hard ? true : false;
     try {
       await systemInterface.interactivelyReloadModule(
-        null, m.name, reloadDeps, resetEnv);      
+        null, m.name, reloadDeps, resetEnv);
       await this.selectModuleNamed(m.nameInPackage);
       sourceEditor.scroll = scroll;
       sourceEditor.cursorPosition = cursorPosition;
