+++ ide/js/inspector.js
@@ -3,10 +3,10 @@ import { Color, rect, pt } from "lively.graphics";
 import { obj, arr, promise, string } from "lively.lang";
 import { connect, disconnectAll, signal, disconnect, once } from "lively.bindings";
 import { Morph, DropDownList, HorizontalLayout, morph, CustomLayout, Label, Icon, StyleSheet, config } from "lively.morphic";
-import { Tree, TreeData } from "lively.morphic/components/tree.js";
+import { Tree, TreeData } from "lively.components/tree.js";
 
 import { isBoolean, isString, isNumber } from "lively.lang/object.js";
-import { DropDownSelector, SearchField, LabeledCheckBox } from "../../components/widgets.js";
+import { DropDownSelector, SearchField, LabeledCheckBox } from "lively.components/widgets.js";
 import { last } from "lively.lang/array.js";
 
 import { GridLayout } from "../../layout.js";
@@ -216,7 +216,7 @@ class MorphNode extends InspectionNode {
 
   getSubNode(nodeArgs) {
     let spec = this.propertyInfo[nodeArgs.key] || {};
-    if (nodeArgs.value && nodeArgs.value.isMorph) 
+    if (nodeArgs.value && nodeArgs.value.isMorph)
        return new MorphNode({...nodeArgs, root: this.root})
     return new PropertyNode({
       ...nodeArgs,
@@ -287,7 +287,7 @@ class PropertyNode extends InspectionNode {
         this._propertyWidget.highlight();
       }
     }
-    if (this.renderedNode) 
+    if (this.renderedNode)
       this.renderedNode.width = this._propertyWidget.bounds().width + this.renderedNode.toggle.width + 3;
     if (this.isFoldable) {
       for (let m in this.foldedNodes) {
@@ -418,13 +418,13 @@ class DraggedProp extends Morph {
       }
     }
   }
- 
+
   applyToTarget(evt) {
     let {currentTarget: target, control} = this;
     this.remove();
     MorphHighlighter.removeHighlighters(evt.world);
     if (!target) return;
- 
+
     if (!target.isText || target.editorModeName !== "js") {
       // normal apply prop
       if (control.hasOwnProperty("propertyValue"))
@@ -908,8 +908,8 @@ export default class Inspector extends Morph {
     this.layout.col(0).row(1).group.morph = tree;
 
     this.whenRendered().then(
-      () => { 
-        if (this.targetObject.isMorph && 
+      () => {
+        if (this.targetObject.isMorph &&
             this.targetObject.world() == this.world()) {
           this.targetObject = this.targetObject;
         } else {
@@ -917,10 +917,10 @@ export default class Inspector extends Morph {
         }
       }
     );
-    
+
     super.__after_deserialize__();
   }
-  
+
   __additionally_serialize__(snapshot, ref, pool, addFn) {
     // empty tree
     let submorphs = snapshot.props.submorphs.value;
@@ -929,7 +929,7 @@ export default class Inspector extends Morph {
       if (pool.refForId(id).realObj == this.ui.propertyTree) arr.removeAt(submorphs, i);
     }
   }
-  
+
 
   static get properties() {
     return {
@@ -1097,9 +1097,9 @@ export default class Inspector extends Morph {
     codeEditor.changeEditorMode("js").then(() =>
       codeEditor.evalEnvironment = {
         targetModule: "lively://lively.morphic/inspector",
-        get context() { 
-          return thisBindingSelector.selection == 'selection' ? 
-              codeEditor.owner.selectedObject : codeEditor.owner.targetObject 
+        get context() {
+          return thisBindingSelector.selection == 'selection' ?
+              codeEditor.owner.selectedObject : codeEditor.owner.targetObject
         },
         format: "esm"
       }
@@ -1128,7 +1128,7 @@ export default class Inspector extends Morph {
       return;
     }
     let change = last(this.targetObject.env.changeManager.changesFor(this.targetObject));
-    if (change == this.lastChange && this.lastSubmorphs == printValue(this.targetObject && this.targetObject.submorphs)) 
+    if (change == this.lastChange && this.lastSubmorphs == printValue(this.targetObject && this.targetObject.submorphs))
       return;
     if (this.focusedNode && this.focusedNode.keyString == change.prop) {
       this.repositionOpenWidget();
@@ -1142,7 +1142,7 @@ export default class Inspector extends Morph {
     var v;
     this.originalTreeData && this.originalTreeData.asListWithIndexAndDepth(false).forEach(({node}) => {
       if (node.foldableNode) {
-        v = this.targetObject[node.foldableNode.key][node.key]; 
+        v = this.targetObject[node.foldableNode.key][node.key];
       } else {
         v = this.targetObject[node.key];
       }
@@ -1241,7 +1241,7 @@ export default class Inspector extends Morph {
       {name: "codeEditor", ...textStyle},
       {
         name: 'fix import button', type: "button",
-        fill:  Color.black.withA(.5), 
+        fill:  Color.black.withA(.5),
         fontColor: Color.white,
         borderWidth: 0,
         label: "fix undeclared vars", extent: pt(100, 20)
@@ -1253,7 +1253,7 @@ export default class Inspector extends Morph {
         selection: 'selection',
         items: [{isListItem: true, value: 'target',
                  label: ['this ', null, ...rightArrow, ' target', null]},
-               {isListItem: true, value: 'selection', 
+               {isListItem: true, value: 'selection',
                 label: ['this ', null, ...rightArrow, ' selection', null]}]
       }
     ];
