+++ ide/shell/workspace.js
@@ -2,7 +2,7 @@ import { arr, obj } from "lively.lang";
 import { connect } from "lively.bindings";
 import { pt, Point, Color, Rectangle } from "lively.graphics";
 import { config, morph } from "../../index.js";
-import Window from "../../components/window.js";
+import Window from "lively.components/window.js";
 import ShellEditorPlugin from "./editor-plugin.js";
 import Terminal from "./terminal.js";
