+++ ide/styling/color-palette.js
@@ -3,7 +3,7 @@ import { Color, rect, Complementary, Triadic, Tetradic,
          flatDesignColors, materialDesignColors, webSafeColors } from "lively.graphics";
 import { Morph, CustomLayout, VerticalLayout, HorizontalLayout,
          Text, TilingLayout, Ellipse, Image } from "../../index.js";
-import { DropDownSelector, ModeSelector, Slider } from "../../components/widgets.js";
+import { DropDownSelector, ModeSelector, Slider } from "lively.components/widgets.js";
 import { connect, signal } from "lively.bindings";
 import { num, arr } from "lively.lang";
 import { StyleSheet } from "../../style-sheets.js";
@@ -31,7 +31,7 @@ class ColorPaletteField extends Morph {
       }
     }
   }
-  
+
   onHoverIn() {
     if (!this.color) return;
     const [h,s,b] = this.color.toHSB();
@@ -43,7 +43,7 @@ class ColorPaletteField extends Morph {
   onMouseDown() {
     signal(this, "updateColor", this.color);
   }
-   
+
 }
 
 class ColorHarmonyField extends Morph {
@@ -70,8 +70,8 @@ class ColorHarmonyField extends Morph {
              let [h,s,b] = this.color.toHSB();
              this.submorphs = [
                 new Morph({fill: this.color, extent: pt(80, 50)}),
-                new Text({textString: `${h.toFixed()}, ${s.toFixed(2)}, ${b.toFixed(2)}`, 
-                          fill: Color.transparent, fontColor: Color.gray.darker(), 
+                new Text({textString: `${h.toFixed()}, ${s.toFixed(2)}, ${b.toFixed(2)}`,
+                          fill: Color.transparent, fontColor: Color.gray.darker(),
                           bottomLeft: pt(0,50)})
              ];
            }
@@ -82,7 +82,7 @@ class ColorHarmonyField extends Morph {
   onMouseDown() {
     signal(this, "updateColor", this.color);
   }
-  
+
 }
 
 class HarmonyDisplay extends Morph {
@@ -103,7 +103,7 @@ class HarmonyDisplay extends Morph {
       });
     }
   }
-  
+
 }
 
 class PivotColorControl extends Ellipse {
@@ -171,7 +171,7 @@ class HarmonyVisualizer extends Image {
          pivotControl.center = Point.polar(50 * saturation, angle);
      }
   }
-  
+
 }
 
 class HarmonyPalette extends Morph {
@@ -185,10 +185,10 @@ class HarmonyPalette extends Morph {
         after: ['harmony'],
         initialize() {
           this.submorphs = [
-             new HarmonyDisplay({name: "harmonies"}), 
+             new HarmonyDisplay({name: "harmonies"}),
              this.harmonyControl()]
         }
-      } 
+      }
     }
   }
 
@@ -201,7 +201,7 @@ class HarmonyPalette extends Morph {
     this.get('harmony visualizer').update(colorPalette);
     this.updateInProgress = false;
   }
- 
+
   harmonyControl() {
     var selector,
         controls = new Morph({
@@ -290,8 +290,8 @@ export class ColorPalette extends Morph {
             this.submorphs = [this.fillTypeSelector(), this.paletteView()];
             this.pivotColor = Color.hsb(h,s,1);
             connect(
-              this.get('paletteView'), 'extent', 
-              this.get('fillTypeSelector'), 'width', 
+              this.get('paletteView'), 'extent',
+              this.get('fillTypeSelector'), 'width',
               {converter: ({x}) => x}
             )
             connect(this.get('harmony palette'), 'harmony', this, 'harmony');
@@ -384,7 +384,7 @@ export class ColorPalette extends Morph {
         center: pt(8, 8),
         extent: pt(12, 12)
       }
-    });   
+    });
    }
 
    isHaloItem() { return true }
@@ -402,7 +402,7 @@ export class ColorPalette extends Morph {
       return {
         name: "paletteView",
         submorphs: [
-           this.solidColorPalette(), 
+           this.solidColorPalette(),
            new HarmonyPalette({name: "harmony palette", isLayoutable: false,
                                visible: false, draggable: false})]
       }
