+++ package.json
@@ -11,6 +11,7 @@
     "lively.bindings": "https://github.com/LivelyKernel/lively.bindings",
     "lively.graphics": "https://github.com/LivelyKernel/lively.graphics",
     "lively.halos": "https://github.com/LivelyKernel/lively.halos",
+    "lively.components": "https://github.com/LivelyKernel/lively.components",
     "lively.lang": "^1.0.0",
     "lively.serializer2": "https://github.com/LivelyKernel/lively.serializer2",
     "pepjs": "^0.4.3",
